unify permutive specs
